Remove the subclass implementation. Should make the filters work again!
