+++ service/grails-app/views/entitlement/_externalEntitlement.gson
@@ -1,18 +1,17 @@
-package externalEntitlement;
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.Future
 
-import org.olf.erm.ExternalEntitlement
+import org.olf.erm.Entitlement
 
 import groovy.transform.*
 import groovyx.net.http.HttpException
 
-@Field ExternalEntitlement externalEntitlement
+@Field Entitlement entitlement
 final String objectProperty = 'reference_object'
 def remoteObjValue
-if(externalEntitlement.respondsTo(objectProperty)){
+if(entitlement.respondsTo(objectProperty)){
   try {
-    remoteObjValue = externalEntitlement.invokeMethod(objectProperty, null)
+    remoteObjValue = entitlement.invokeMethod(objectProperty, null)
     if (remoteObjValue instanceof Future) {
       remoteObjValue = (remoteObjValue as Future)?.get()
     }
@@ -28,15 +27,15 @@ if(externalEntitlement.respondsTo(objectProperty)){
 }
 
 json {
-  'id' externalEntitlement.id
+  'id' entitlement.id
   'type' 'external'
-  'authority' externalEntitlement.getAuthority()
-  'reference' externalEntitlement.getReference()  
-  'explanation' externalEntitlement.getExplanation()
-  'startDate' externalEntitlement.getActiveFrom()
-  'endDate' externalEntitlement.getActiveTo()
-  'contentUpdated' externalEntitlement.getContentUpdated()
-  'haveAccess' externalEntitlement.getHaveAccess()
+  'authority' entitlement.getAuthority()
+  'reference' entitlement.getReference()  
+  'explanation' entitlement.getExplanation()
+  'startDate' entitlement.getActiveFrom()
+  'endDate' entitlement.getActiveTo()
+  'contentUpdated' entitlement.getContentUpdated()
+  'haveAccess' entitlement.getHaveAccess()
   
   if (remoteObjValue) {
     call( objectProperty, remoteObjValue )
